Nightmode (screensaver) persistence through screen off/on.
WindowManager layoutparams didn't allow screensaver to
stay on through screen on/off toggles similar to the
funcitonality in 4.1.1 desk clock. Also onPause() in
ScreensaverActivity.java executed finish() which stops
the screensaver.

I commented out the finish(), and put the setWakeLock()
method back to what it was in the 4.1.1 desk clock.

Change-Id: If39ab2b3ee6dbb9ed1afcd9c46cbe1d41aab8706
Signed-off-by: Joseph Stover <doduff@gmail.com>