Update patch set 2
Patch Set 2: Code-Review+1

I downloaded the change, built the DeskClock (full_toro-userdebug, but don't think it should matter) and tested it on the device with 4.1.1 and in a 4.2.2 emulator. I had to change the namespace (com.android.deskclock and com.android.alarmclock in all the files) in order to have it install due to presence of the stock deskclock (didn't want to build an entire system image). It functions as expected according to the changes: pressing the power button to turn the screen off/on or letting the screen timeout and turn off leaves the screensaver intact. 

There is a behavior that is present in the 4.1.1 clock as well. When turning the screen back on, the soft keys sometimes flash (In the emulator, the red border flashes). I don't know if this is a potential problem, and possibly there is a reason someone stopped the nightmode persistence. For my device (toro), nightmode persistence was not present in 4.0.2, then was in 4.0.4 and 4.1.1, but was gone again in 4.2.2. So I'm not sure what most people want, but from the looks of the code, there is the desire to have nightmode persist (see parameter mFlags). However, setWakeLock() only implemented the persistence while plugged in, and onPause() completely prevented it with a finish().

Patch-set: 2