Update patch set 1
Patch Set 1:

Translations are done by a different system, so I need to take your translation and ask our translation people to approve it,then it will be imported into the Android code.
But this specific fix was done some time ago and will be visible with the next AOSP push.

Patch-set: 1