If a Snoozed alarm is disabled, remove snoozed notification
If the user snoozes an alarm and then goes back into the Clock app
and manually disabled the alarm, the notification is still enabled
and keeps triggering. This change will now remove any notification
when an alarm is updated, then it will reschedule all alarms that
are outstanding (as normal).

Issue: https://code.google.com/p/android/issues/detail?id=57583

Change-Id: I70734227d0f675340d6be9a4dfb0d9c0e47b19e3
Signed-off-by: howardb <android@howardb.com>