If a Snoozed alarm is disabled, remove snoozed notification
If an alarm is snoozed and the user opens the Clock app and
manually disabled that alarm, the notification remains and
continues to trigger. This change will remove any snoozed
notifications if an alarm is updated. The existing code will
then reschedule any alarms, if need be.

Issue: https://code.google.com/p/android/issues/detail?id=57583

Change-Id: I204068bbe99da34597ff9bc272d11c068bd4198d
Signed-off-by: howardb <android@howardb.com>