Update patch set 3
Patch Set 3:

On second thought, the way it is implemented right now will cause any activity alias that ends with AlarmClock to trigger the new code.
I think a better way is to create a small activity that will start the DeskClock activity with the alarms tab as an extra.

Patch-set: 3
Reviewer: Gerrit User 1010105 <1010105@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Verified=0