Update patch set 1
Patch Set 1:

Hi, thanks for identifying this bug!

We have a different implementation of the fix to this issue. Would you be willing to wait for our upstream fix, which will come in a later release? Or is this urgent for you?

Patch-set: 1
Reviewer: Gerrit User 1060503 <1060503@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Verified=0