Fix alarm notification issue
There is below error when alarm time trigged.
You can't see any UI you can click to close the alarm when the alarm is trigged.
If you close it in destclock and set a new alarm, alarm may no be trigged when alarm time got.
Error log:
 E NotificationService: No Channel found for pkg=com.android.deskclock, channelId=null,
 id=3, tag=null, opPkg=com.android.deskclock, callingUid=10043, userId=0,
 incomingUserId=0, notificationUid=10043, notification=Notification

When the target is Android O, we must implement one or more notification channels to
display notifications to users.
We create alarm default channel when ACTION_BOOT_COMPLETED in AlarmInitReceiver.
We need add setChannelId when create the notification in the destclock apk.

Change-Id: Ifc07b403034b232530f7f39ef9f5d17c3dd63342
Signed-off-by: zhang sanshan <sanshan.zhang@nxp.com>