Update patch set 1
Patch Set 1:

(1 comment)

This change is much closer to what the Play Store version of this app does, which is good. Disabling sound, vibrate etc. like this change does is correct.

See my comment though. Was there a particular reason for using IMPORTANCE_MAX or was it due to the naming confusion, which would certainly be understandable?

Patch-set: 1
CC: Gerrit User 1005571 <1005571@85c56323-6fa9-3386-8a01-6480fb634889>